+++ src/main/java/com/school21/euphoria/repository/CarRepository.java
@@ -4,9 +4,7 @@ import com.school21.euphoria.entity.CarEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface CarRepository extends JpaRepository<CarEntity, Long> {
+public interface CarRepository extends JpaRepository<CarEntity, Integer> {
     void deleteById(Long id);
 }
